feat: Add GitHub Actions for .env generation and CI workflow
- Introduced a new action to generate a .env file using secrets.
- Created a CI workflow for Python package with steps for setup, linting, testing, and Docker service management.
- Updated .gitignore to exclude .env files from version control.
- Enhanced Makefile with new targets for testing and service management.
- Added scripts for environment variable checks and encoding .env files.
- Implemented integration tests for Docker services, network accessibility, and volume management.
- Removed outdated test files and organized tests by type (unit, integration, etc.).
- Updated project version to 1.2.8 in package.json and sonar-project.properties.
