+++ pytest.ini
@@ -2,8 +2,13 @@
 minversion = 6.0
 addopts = -ra -q
 testpaths = src/tests
+
 markers =
-    network: Testes relacionados à rede
-    services: Testes relacionados aos serviços
-    integration: Testes de integração
-    unit: Testes unitários
+    unit: Testes unitários que não dependem de containers Docker.
+    integration: Testes de integração que exigem containers ativos.
+    network: Testes relacionados à rede Docker e resolução DNS.
+    services: Testes que verificam se os serviços estão rodando e nas portas corretas.
+    docker: Testes que utilizam docker-compose para ambientes reais.
+    testcontainers: Testes que usam a lib testcontainers para criar ambientes isolados.
+    volumes: Testes relacionados à criação e montagem de volumes Docker.
+    dns: Testes relacionados à resolução de DNS entre containers.
