+++ scripts/run-ci-local.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+act push -P ubuntu-latest=ghcr.io/catthehacker/ubuntu:act-latest \
+  --container-options "-v $(pwd)/.env:/home/runner/work/infra-devtools/infra-devtools/.env" \
+  -s GITHUB_TOKEN=$ACT_TOKEN
