+++ sonar-project.properties
@@ -1,6 +1,6 @@
 sonar.projectName=Infra Dev Tools
 sonar.projectKey=infra-devtools
-sonar.projectVersion=1.2.7
+sonar.projectVersion=1.2.8
 
 sonar.host.url=${SONAR_HOST_URL}
 sonar.token=${SONAR_TOKEN_INFRA_DEVTOOLS}
