+++ src/tests/integration/test_network_docker_services.py
@@ -4,6 +4,7 @@ import logging
 
 logging.basicConfig(level=logging.INFO)
 
+# Lista de containers e respectivas portas para validação de acessibilidade
 containers = [
     ("infra-default-mongo", 27017),
     ("infra-default-mongo-express", 8081),
@@ -24,16 +25,22 @@ containers = [
     ("infra-default-mongodb-exporter", 9216),
     ("infra-default-postgres-exporter", 9187),
 ]
+
+@pytest.mark.integration
 @pytest.mark.network
 def test_network_exists(host):
-    """Verifica se a rede Docker principal existe."""
+    """🕸️ Verifica se a rede Docker principal 'infra-default-shared-net' existe."""
     network_name = "infra-default-shared-net"
-    networks = host.check_output("docker network ls --filter name=%s --format '{{.Name}}'", network_name)
-    assert network_name in networks, f"A rede {network_name} não foi encontrada!"
+    networks = host.check_output(
+        "docker network ls --filter name=%s --format '{{.Name}}'", network_name
+    )
+    assert network_name in networks, f"❌ A rede {network_name} não foi encontrada!"
 
-
+@pytest.mark.integration
+@pytest.mark.network
 @pytest.mark.parametrize("container_name, port", containers)
 def test_container_accessible(host, container_name, port):
+    """🔌 Verifica se os containers estão acessíveis via localhost e porta mapeada."""
     connected = False
     try:
         with socket.create_connection(("localhost", port), timeout=5):
@@ -42,9 +49,10 @@ def test_container_accessible(host, container_name, port):
         print(f"Erro ao conectar com {container_name}:{port} -> {e}")
     assert connected, f"❌ Falha ao conectar em {container_name}:{port}"
 
+@pytest.mark.integration
 @pytest.mark.network
 def test_dns_resolution_getent(host):
-    """Verifica resolução de DNS usando getent hosts"""
+    """🌐 Verifica se os containers resolvem DNS via getent hosts."""
     services = [
         "infra-default-mongo",
         "infra-default-redis",
@@ -62,13 +70,12 @@ def test_dns_resolution_getent(host):
 
     for service in services:
         result = host.run(f"docker exec {service} getent hosts {target_host}")
-        assert result.rc == 0 and target_host in result.stdout, f"[getent] DNS FAIL: {service} -> {target_host}"
+        assert result.rc == 0 and target_host in result.stdout, f"❌ [getent] DNS FAIL: {service} → {target_host}"
 
-
+@pytest.mark.integration
 @pytest.mark.network
 def test_service_dns_resolution(host):
-    """Verifica se os serviços podem resolver DNS antes de conectar."""
-    
+    """📡 Verifica se os serviços resolvem seu próprio DNS antes de conectar."""
     services = [
         "infra-default-mysql",
         "infra-default-postgres",
@@ -85,6 +92,5 @@ def test_service_dns_resolution(host):
                 logging.error(f"❌ Falha na resolução DNS para {service}")
                 pytest.fail(f"❌ Serviço {service} não conseguiu resolver DNS corretamente")
         except Exception as e:
-                logging.error(f"❌ Erro ao testar DNS para {service} -> {e}")
-                pytest.fail(f"❌ Erro crítico na verificação DNS para {service}: {e}")
-
+            logging.error(f"❌ Erro ao testar DNS para {service} → {e}")
+            pytest.fail(f"❌ Erro crítico na verificação DNS para {service}: {e}")
